requirements.txt
streamlit
sqlalchemy
snowflake-sqlalchemy
psycopg2-binary
pandas
numpy
python-dateutil
scikit-learn